feat: 파일 합친 후 오류 수정
모든 브랜치 파일 합친 후 실행에 생긴 오류 수정, SaveTripService에서 TripForJson으로 다운캐스팅 하는 부분이 필요함.
