+++ toy_project1/build.gradle
@@ -12,8 +12,8 @@ repositories {
 dependencies {
     testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
-    compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.24'
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.15.2'
+    implementation group: 'com.opencsv', name: 'opencsv', version: '5.7.1'
 }
 
 test {
